requirements.txt
openai
pandas
numpy
streamlit
ollama
python-dotenv